# docs/swagger.yaml
basePath: /
definitions:
...
  handlers.genreRequest:
    properties:
      name:
        type: string
    type: object
  handlers.updatePasswordRequest:
    properties:
      new_password:
        type: string
      old_password:
        type: string
    type: object
  handlers.userUpdate:
    properties:
      born_at:
        type: string
      name:
        type: string
      phone_number:
        type: string
    type: object
...
  models.Genre:
    properties:
      id:
        type: integer
      name:
        type: string
    type: object
...
  models.User:
    properties:
      born_at:
        type: string
      email:
        type: string
      id:
        type: integer
      name:
        type: string
      phone_number:
        type: string
      role:
        type: string
    type: object
host: localhost:8080
info:
  contact: {}
  description: REST API для приложения Özinshe.
  title: Özinshe API
  version: "1.0"
paths:
...
  /genres:
    get:
      description: Возвращает списко всех доступных жанров.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Genre'
            type: array
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Получить список всех жанров
      tags:
      - Genres
    post:
      consumes:
      - application/json
      description: Создает новый жанр. Доступно только администраторам.
      parameters:
      - description: Данные для создания
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handlers.genreRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.Genre'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Создает жанр
      tags:
      - Genres
  /genres/{id}:
    delete:
      description: Удаляет выбранный жанр. Доступно только администраторам.
      parameters:
      - description: ID жанра
        in: path
        name: id
        required: true
        type: integer
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Удалить жанр
      tags:
      - Genres
    get:
      description: Возвращает выбранный по ID жанр
      parameters:
      - description: ID жанра
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Genre'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Получить жанр по ID
      tags:
      - Genres
    patch:
      consumes:
      - application/json
      description: Обновляет выбранный жанр. Доступно только администраторам.
      parameters:
      - description: ID жанра
        in: path
        name: id
        required: true
        type: integer
      - description: Данные для обновления
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handlers.genreRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Genre'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Обновить жанр
      tags:
      - Genres
...
  /users/me:
    get:
      description: Возвращает данные пользователя, авторизованного через JWT.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.User'
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Получить текщего пользователя
      tags:
      - Users
    patch:
      consumes:
      - application/json
      description: Обновляет имя, номер телефона и дату рождения текущего пользователя.
      parameters:
      - description: Данные для обновления
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handlers.userUpdate'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.User'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Обновить данные текущего пользователя
      tags:
      - Users
  /users/me/password:
    patch:
      consumes:
      - application/json
      description: Обновляет пароль текущего пользователя.
      parameters:
      - description: Данные для обновления
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handlers.updatePasswordRequest'
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Обновить пароль текущего пользователя
      tags:
      - Users
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"